+++ versiontools/__init__.py
@@ -118,6 +118,9 @@ class Version(tuple):
         if releaselevel not in ('dev', 'alpha', 'beta', 'candidate', 'final'):
             raise ValueError(
                 "releaselevel %r is not permitted" % (releaselevel,))
+        if releaselevel in ('alpha', 'beta', 'candidate') and serial == 0:
+            raise ValueError(
+                "serial must be greater than zero for %s releases" % releaselevel)
         obj = tuple.__new__(cls, (major, minor, micro, releaselevel, serial))
         object.__setattr__(obj, '_source_tree', cls._find_source_tree())
         object.__setattr__(obj, '_vcs', None)